Refactored malpolon to better separate datasets & datamodules from the experiment scripts. Also brought small fixes to existing classes. - Example files: moved as much as possible the datamodules to their corresponding 'malpolon.data.dataset' file. - Created 'malpolon.data.dataset.torchgeo_sentinel2.py' - Modified example config files to match their respective experiments. - Separated RasterSentinel2 from RasterTorchGeoDataset - Modified 'malpolon/data/datasets/__init__.py' to correctly import sentinel2 and raster_torchgeo classes - geolifeclef2022.py : added more flexibility in auto download method. It is now possible to add 'geolifeclef-2022-lifeclef-2022-fgvc9' a the end of the dataset_path in the config file and the download method will take that into account and ajust 'self.root' accordingly. - geolifeclef2022.py : added download argument in 'MiniGeoLifeCLEF2022Dataset' init method. - Moved metrics constant used by examples to read their config files' metrics arguments, to 'malpolon/data/utils.py' - 'standard_prediction_systems.py' : added task and metrics hparams processing under 'FineTuningClassificationSystem', and renamed the class 'ClassificationSystem'. Also removed redfinitions of 'FineTuningClassificationSystem' (named 'ClassificationSystem') from example files. - torchgeo_datasets.py : added docstrings and cleaned file - Example files : moved as much transform classes to their 'transform.py' associated file.
