+++ examples/ecologists/micro_geolifeclef2022/cnn_on_rgb_patches.py
@@ -3,9 +3,7 @@ from __future__ import annotations
 import os
 
 import hydra
-import omegaconf
 import pytorch_lightning as pl
-import torchmetrics.functional as Fmetrics
 from omegaconf import DictConfig
 from pytorch_lightning.callbacks import ModelCheckpoint
 from torchvision import transforms
@@ -14,11 +12,7 @@ from transforms import RGBDataTransform
 from malpolon.data.data_module import BaseDataModule
 from malpolon.data.datasets.geolifeclef2022 import MicroGeoLifeCLEF2022Dataset
 from malpolon.logging import Summary
-from malpolon.models import FinetuningClassificationSystem
-
-FMETRICS_CALLABLES = {'binary_accuracy': Fmetrics.classification.binary_accuracy,
-                      'multiclass_accuracy': Fmetrics.classification.multiclass_accuracy,
-                      'multilabel_accuracy': Fmetrics.classification.multilabel_accuracy, }
+from malpolon.models import ClassificationSystem
 
 
 class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
@@ -38,9 +32,11 @@ class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
         train_batch_size: int = 32,
         inference_batch_size: int = 256,
         num_workers: int = 8,
+        download: bool = True,
     ):
         super().__init__(train_batch_size, inference_batch_size, num_workers)
         self.dataset_path = dataset_path
+        self.download = download
 
     @property
     def train_transform(self):
@@ -52,7 +48,8 @@ class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
                 transforms.RandomHorizontalFlip(),
                 transforms.RandomVerticalFlip(),
                 transforms.Normalize(
-                    mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]
+                    mean=[0.485, 0.456, 0.406],
+                    std=[0.229, 0.224, 0.225]
                 ),
             ]
         )
@@ -64,7 +61,8 @@ class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
                 lambda data: RGBDataTransform()(data["rgb"]),
                 transforms.CenterCrop(size=224),
                 transforms.Normalize(
-                    mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]
+                    mean=[0.485, 0.456, 0.406],
+                    std=[0.229, 0.224, 0.225]
                 ),
             ]
         )
@@ -74,7 +72,7 @@ class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
             self.dataset_path,
             subset="train",
             use_rasters=False,
-            download=True,
+            download=self.download,
         )
 
     def get_dataset(self, split, transform, **kwargs):
@@ -87,80 +85,6 @@ class MicroGeoLifeCLEF2022DataModule(BaseDataModule):
             **kwargs
         )
         return dataset
-
-
-class ClassificationSystem(FinetuningClassificationSystem):
-    """Classification task class."""
-    def __init__(
-        self,
-        model: dict,
-        lr: float,
-        weight_decay: float,
-        momentum: float,
-        nesterov: bool,
-        metrics: dict,
-        task: str = 'classification_multiclass',
-        hparams_preprocess: bool = True,
-    ):
-        """Class constructor.
-
-        Parameters
-        ----------
-        model : dict
-            _description_
-        lr : float
-            learning rate
-        weight_decay : float
-            weight decay
-        momentum : float
-            value of momentum
-        nesterov : bool
-            if True, uses Nesterov's momentum
-        metrics : dict
-            dictionnary containing the metrics to compute.
-            Keys must match metrics' names and have a subkey with each
-            metric's functional methods as value. This subkey is either
-            created from the FMETRICS_CALLABLES constant or supplied,
-            by the user directly.
-        task : str, optional
-            machine learning task (used to format labels accordingly),
-            by default 'classification_multiclass'
-        hparams_preprocess : bool, optional
-            if True performs preprocessing operations on the hyperparameters,
-            by default True
-        """
-        if hparams_preprocess:
-            task = task.split('classification_')[1]
-            try:
-                metrics = omegaconf.OmegaConf.to_container(metrics)
-                for k, v in metrics.items():
-                    if 'callable' in v:
-                        metrics[k]['callable'] = eval(v['callable'])
-                    else:
-                        metrics[k]['callable'] = FMETRICS_CALLABLES[k]
-            except ValueError as e:
-                print('\n[WARNING]: Please make sure you have registered'
-                      ' a dict-like value to your "metrics" key in your'
-                      ' config file. Defaulting metrics to None.\n')
-                print(e, '\n')
-                metrics = None
-            except KeyError as e:
-                print('\n[WARNING]: Please make sure the name of your metrics'
-                      ' registered in your config file match an entry'
-                      ' in constant FMETRICS_CALLABLES.'
-                      ' Defaulting metrics to None.\n')
-                print(e, '\n')
-                metrics = None
-
-        super().__init__(
-            model,
-            lr,
-            weight_decay,
-            momentum,
-            nesterov,
-            metrics,
-            task,
-        )
 
 
 @hydra.main(version_base="1.1", config_path="config", config_name="cnn_on_rgb_patches_config")
