+++ malpolon/data/datasets/__init__.py
@@ -1,6 +1,6 @@
 from .geolifeclef2022 import GeoLifeCLEF2022Dataset, MiniGeoLifeCLEF2022Dataset
-from .torchgeo_datasets import RasterTorchGeoDataset, RasterSentinel2
-
+from .torchgeo_datasets import RasterTorchGeoDataset
+from .torchgeo_sentinel2 import RasterSentinel2
 
 __all__ = [
     "GeoLifeCLEF2022Dataset",
